# backend/docs/swagger.yaml
basePath: /
definitions:
...
  httperrors.ErrorResponse:
    properties:
      details:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
  repository.Chat:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      metadata:
        items:
          type: integer
        type: array
      title:
        type: string
      type:
        $ref: '#/definitions/repository.ChatType'
      updated_at:
        type: string
    type: object
...
  repository.Message:
    properties:
      chat_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      reply_message_id:
        type: string
      sender_id:
        type: string
      topic_id:
        type: string
      updated_at:
        type: string
    type: object
...
  repository.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      last_seen_at:
        type: string
      metadata:
        items:
          type: integer
        type: array
      password_hash:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  services.ChatWithLastMessage:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      last_message:
        $ref: '#/definitions/repository.Message'
      metadata:
        items:
          type: integer
        type: array
      title:
        type: string
      type:
        $ref: '#/definitions/repository.ChatType'
      updated_at:
        type: string
    type: object
...
  services.InvitationResponse:
    properties:
      id:
        type: string
    type: object
...
  services.UpdateChatDTO:
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for a messenger application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Messenger API
  version: "1.0"
paths:
...
  /chats/:
    get:
      consumes:
      - application/json
      description: Returns a list of all chats with their most recent message.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.ChatWithLastMessage'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all chats
      tags:
      - chats
  /chats/{id}:
    get:
      consumes:
      - application/json
      description: Returns a chat by ID.
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repository.Chat'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a chat by ID
      tags:
      - chats
    put:
      consumes:
      - application/json
      description: Update a chat's title and metadata.
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: string
      - description: Chat details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.UpdateChatDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repository.Chat'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update chat
      tags:
      - chats
  /chats/{id}/avatar:
    post:
      consumes:
      - application/json
      description: Upload an avatar image for a chat.
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repository.Chat'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload chat avatar
      tags:
      - chats
  /chats/{id}/invitation:
    post:
      consumes:
      - application/json
      description: Create an invitation link for a chat. Returns the invitation ID.
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.InvitationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create invitation
      tags:
      - chats
  /chats/{id}/invite/{user_id}:
    post:
      consumes:
      - application/json
      description: Add a user directly to a chat by user ID.
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Invite user to chat
      tags:
      - chats
  /chats/{id}/leave:
    post:
      consumes:
      - application/json
      description: Remove the authenticated user from a chat.
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Leave chat
      tags:
      - chats
  /chats/{id}/members:
    get:
      consumes:
      - application/json
      description: Returns all members of a chat.
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repository.User'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List chat members
      tags:
      - chats
  /chats/{id}/messages:
    get:
      consumes:
      - application/json
      description: Returns a list of all messages in a chat.
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repository.Message'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all messages in a chat
      tags:
      - messages
    post:
      consumes:
      - application/json
      description: Creates a new message in a chat.
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: string
      - description: Message details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.CreateMessageDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repository.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new message in a chat
      tags:
      - messages
  /chats/{id}/mute:
    put:
      consumes:
      - application/json
      description: Mute or unmute notifications for a chat.
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: string
      - description: Mute status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.MuteChatDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repository.Chat'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mute/unmute chat
      tags:
      - chats
  /chats/{id}/my-role:
    get:
      consumes:
      - application/json
      description: Returns the role of the authenticated user in a chat.
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get my role in chat
      tags:
      - chats
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by your JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"